Ensure tmp dir is deleted on exit
Deleting the tmp dir with a command in the script leaves the potential
for the script to exit before reaching that line, leaving the tmp dir
undeleted. Further, deleting the tmp dir with "rmdir" will only work
if the tmp dir is empty which may be the case if the script failed
to move some of the files during installation. This change addresses
both these issues.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>