Store the current date in a text file
This gives an easy way to determine the latest date on which a
successful build occurred. The next step will be reading the date file
in the install script so it always installs the latest version of dune.
That will fix the issue where people in eastern timezones can't access
the nightly builds before a certain time each day, and also it will mean
that if the nightly build fails, the install script will install the
last version of the binary distribution where the build succeeded.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>