Use posix args to `date` in github action
This fixes a bug where the build time is missing from version numbers on
macOS. The build time is missing because the `date` command on macOS is
not the gnu implementation, so lacks the `--iso-8601` argument, causing
the command to fail. The fix is to explicitly describe a date format
which is very close to ISO8601, the one difference being that it uses a
"Z" to denote the UTC timezone.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>