Run env scripts from external files
This fixes an issue where installing dune multiple times results in
multiple additions to the shell rc file. The most naive way to achive
this would just be to not modify the shell rc if it looks like dune has
already been installed. The problem with that approach is that we may
want to update the env changes in the future. Moving all env changes to
a separate file sourcef from the shell rc gives us the ability to change
the contents of that file in future updates without needing to modify
the shell rc.

It also simplifies the install script as we no longer need to put shell
commands in bash strings and deal with the complexities of escaping
characters correctly.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>