Add ocamllsp wrapper script
This script runs the ocamllsp dev tool in an environment where
the ocamlformat dev tool is in PATH, first making sure that the
ocamlformat and ocamllsp dev tools are installed. If the script is run
anywhere other than a dune project with a lockdir then it falls back to
running `ocamllsp` from PATH.

This change doesn't install the `ocamllsp` script yet, it just adds it
to the archive. Installing the script will come in a later change to
make sure we don't update the install script before at least one archive
contains the wrapper script.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>