Use `--no-same-user` always (if possible)
By default tar unpacks with the current user permissions. But when
running as root (e.g. Docker) then tar will unpack with the owner set to
the owner from the tar archive instead of the user. This leads to files
with strange owners and git refusing to touch the cached repository.

Signed-off-by: Marek Kubica <marek@tarides.com>