Install to `.local`
This changes the location into which Dune is installed into the users
`.local` folder, which is fairly standard for a lot of user-installed
packages.

It also changes the locations for env & completions to fit with the
conventions and reworks the config editing to compute the paths.

Finally it adds a fallback for when SHELL isn't set. Dash doesn't set
SHELL so it prints an error (and continues with the * branch). This
makes avoids the error message.

Signed-off-by: Marek Kubica <marek@tarides.com>