Fix issue with paste on macOS
A recent update to macOS appears to have changed the paste command so
that it no longer assumes its inputs is stdin when no input file is
specified. It does still follow the convention of interpreting an input
file named "-" as stdin, so the fix is to pass "-" as the input file to
paste.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>